requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Login>=0.6,<1.0
Flask-WTF>=1.2,<2.0
email-validator>=2.1,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
discord.py>=2.4,<3.0
waitress>=3.0,<4.0
